Parse appId and baseUrl into AppConfig struct
